requirements.txt
aniso8601==8.0.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dotenv==0.14.0
pytz==2020.1
requests==2.25.1
six==1.15.0
urllib3==1.26.5
Werkzeug==1.0.1
